requirements/testing.txt
-r optional.txt
coveralls
django-environ
enum34 ; python_version < '3.4'
factory-boy
mock
pathlib ; python_version < '3.3'
psycopg2-binary
pytest==3.1.3 ; python_version < '3.4'
pytest ; python_version > '3.3'
pytest-cov
pytest-django>=3.4.3
# pytest-djangoapp
pytest-doctest-ellipsis-markers
pytest-html
pytest-lazy-fixture
pytest-mock
pytest-splinter
pytest-sugar
pytest-xdist
splinter
